### 2023-04-12 #5357 @editorialbot
[PRE REVIEW]: qujax: Simulating quantum circuits with JAX
**Submitting author:** @SamDuffield (<a href="http://orcid.org/0000-0002-8656-8734">Samuel Duffield</a>)
**Repository:** https://github.com/CQCL/qujax
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.4
