### 2023-05-18 #5357 @lucydot
Hi @kyleniemeyer, it looks interesting! I don't have experience in quantum computation or JAX, though I suspect we may not have any editors who are. Comfortable with Quant Mech more generally and Python. I'm happy to take it on if there isn't a better suited editor.
