### 2023-05-24 #5357 @amitkumarj441
> @meandmytram @jmiszczak @amitkumarj441 - Hello 👋
> Are you able to review this submission for JOSS? It it Python 🐍 + Quantum Computing :electron:

Hi @lucydot, I am in for the reviewing this submission.
