### 2023-05-24 #5357 @SamDuffield
> Hello @SamDuffield - thanks for your submission to JOSS, and for your patience!
>
> Spotted a small typo when reading through the paper `The core difference is that qujax is designed to purely functional.` --> `The core difference is that qujax is designed to be purely functional.`

No problem, thanks for picking this up! Good catch on the typo, now amended
