### 2023-05-31 #5504 @lucydot
@SamDuffield I can see that the paper is already exceeding the (not completely strict) word limit of ~1000 words. Please bear this in mind when making amendments - if you want to add something to the paper, something else may need to come out. I suggest prioritising (keeping in) your high level summary at the start of the paper, and the statement of need section. The tensor descriptions I think are fine for the paper at its current length.
