### 2023-06-28 #5504 @lucydot
@jmiszczak If you are happy that each point in the checklist is met, then yes go ahead and check each point off. A brief description of what you have tested (e.g. worked through the tutorials) would be useful, and anything that your review has raised that is not an acceptance-blocker, but would improve the quality of the software, is encouraged.

@meandmytram, a reminder for your review.
