### 2023-06-29 #5504 @jmiszczak
Package installation is handled via pip. Basic examples with variational quantum algorithms are provided in notebooks. To run notebooks, one has to additionally instal Jupyter.

The paper provides a good explanation of the provided functionality. Delivered documentation and usage examples are sufficient to start working with the package.

The major advantage of qujax is the ability to use just-in-time compilation. Another nice feature is the functional approach. These two are the major features distinguishing it among other packages.

I highly recommend publication in JOSS.
