### 2023-07-09 #5504 @meandmytram
Hey everyone, just dropping my quick review here.

I was able to install the package through the use of pip and play with all the provided tutorials as well as digging a bit through the API. I found the paper well written and the statement of need the state clearly justifying the fact that the package fills in a specific niche. In addition, the paper offers a comprehensive elucidation of the available functionality. The provided documentation and usage examples are make starting work with the package easy. I totally agree with @jmiszczak on the point of the two distinguishing features of qujax and I do think that it fulfils the need of the community in this area.

Would be cool to see its comparison with QC simulation packages using other backends, but this can be left as a direction for future research.

I endorse the publication of this work in JOSS.
