+++ .github/workflows/docs.yml
@@ -32,6 +32,6 @@ jobs:
         cd .github/workflows/docs        
         sphinx-build ../../../docs . -a
     - name: Upload docs as artefact
-      uses: actions/upload-pages-artifact@v1
+      uses: actions/upload-pages-artifact@v2
       with:
         path: .github/workflows/docs
