### 2023-08-10 d5fe54e
Update paper.md
    - correct formatting of reference list on line 68
    - change to reference of the book (this will need updating with the page number)
