### 2023-08-10 #5504 @lucydot
Hi @SamDuffield -

Just reading through the paper now.

- I can see three arxiv references from a year+ ago. Can you confirm that these have not been published elsewhere in the interim? From a quick search I cannot find anything.
- I can't find the book that is referenced (Nielson and Chuang) with the provided information. I could find a text from 2000 (reprinted 2010) from Cambridge University Press. Please could you double check this reference.
- I've issued a PR with two changes: https://github.com/CQCL/qujax/pull/85
    - correct formatting of reference list on line 68
    - change to reference of the book (this will need updating with the page number)
- please run `@editorialbot generate pdf` to check formatting after merging PR
