### 2023-08-11 #5504 @SamDuffield
@lucydot thanks for checking through!

Hopefully the references are all ok now (I have checked and don't believe the arxiv references have been published elsewhere).

We'll look into doing the release and archival early next week
