### 2023-08-21 db99a9f
Remove e.g. and outer bracket in reference list
Previously there were double brackets around the reference list. The e.g. is now implicit, which I think is acceptable - as I do not think people would assume reference lists are complete unless indicated. Will need re-phrasing if not the case.
