### 2023-08-21 #5504 @lucydot
@SamDuffield the formatting of the reference list is not quite right - another MR (please check if it looks correct after merging ) - https://github.com/CQCL/qujax/pull/91
