### 2023-09-05 8e04b14
Remove JAX ecosystem references from paper
