### 2023-09-05 #5504 @gamatos
Hi @lucydot, just updated and merged your pull request https://github.com/CQCL/qujax/pull/91!
