### 2023-09-11 a372e04
Improve base README.md
- Write high-level description of qujax
- Introduce table of contents
- Introduce Quick start section
- Move description of statetensors/densitytensors to Quick start and add more details
- Add a few more package usage details in Quick start
- Update citation with upcoming JOSS paper
- Merge list of example notebooks into the base folder README.md
- Delete README.md in examples folder
- Minor rewriting and rephrasing
