+++ README.md
@@ -1,36 +1,50 @@
 # qujax
 
-Represent a (parameterised) quantum circuit as a pure [JAX](https://github.com/google/jax) function that
-takes as input any parameters of the circuit and outputs either a _statetensor_ or a _densitytensor_ depending on
-the choice of simulator.
-- The statetensor encodes all $2^N$ amplitudes of the quantum state in a tensor version
-of the statevector, for $N$ qubits.
-- The densitytensor represents a tensor version of the
-$2^N \times 2^N$ density matrix (allowing for mixed states and generic Kraus operators).
+* [Installation](#installation)
+* [Quick start](#quick-start)
+  + [Pure state simulation](#pure-state-simulations)
+  + [Mixed state simulation](#mixed-state-simulations)
+* [Converting from TKET](#converting-from-tket)
+* [Examples](#examples)
+* [Contributing](#contributing)
+* [Citing qujax](#citing-qujax)
+* [API Reference](https://cqcl.github.io/qujax/api/)
 
-Either representation can then be used downstream for exact expectations, gradients or sampling. A JAX implementation
-of a quantum circuit is useful for runtime speedups, automatic differentiation, support for GPUs/TPUs and compatibility
-with other JAX code and packages.
+qujax is a [JAX](https://github.com/google/jax)-based Python library for the classical simulation of gate-based quantum computation. It is designed to be *simple*, *fast* and *flexible*.
 
-Some useful links:
-- [Documentation](https://cqcl.github.io/qujax/api/)
-- [PyPI](https://pypi.org/project/qujax/)
-- [Example notebooks](https://github.com/CQCL/qujax/tree/main/examples)
-- [pytket-qujax](https://github.com/CQCL/pytket-qujax)
+It follows a functional programming design by translating quantum circuits into pure functions. This allows qujax to [seamlessly and directly interface with JAX](https://jax.readthedocs.io/en/latest/notebooks/Common_Gotchas_in_JAX.html#pure-functions), enabling access to its powerful automatic differentiation tools, just-in-time compiler, vectorization capabilities, GPU/TPU integration and growing ecosystem of packages.
+
+qujax can be used for both pure and mixed quantum state simulation. It supports both the standard gate set and user-defined custom operations, including general quantum channels, allowing one to model device noise and errors. 
+
+An overview of the core functionalities of qujax can be found in [Quick start](#quick-start) section. More advanced use-cases, including the training of parameterised quantum circuits, are listed in [Examples](#examples).
 
 
-## Install
+## Installation
+
+qujax is [hosted on PyPI](https://pypi.org/project/qujax/) and can be installed via the pip package manager
 ```
 pip install qujax
 ```
 
-## Statetensor simulations with qujax
+## Quick start
+
+**Important note: qujax circuit parameters are expressed in units of $\pi$ (e.g. in the range $[0,2]$ as opposed to $[0, 2\pi]$)**.
+
+### Pure state simulation
+
+We start by defining the quantum gates making up the circuit, along with the qubits they act on and the indices of the parameters for each gate. 
+
+A list of all gates can be found [here](https://github.com/CQCL/qujax/blob/main/qujax/gates.py) (custom operations can be included by [passing an array or function](https://cqcl.github.io/qujax/api/get_params_to_statetensor_func.html) instead of a string).
+
 ```python
 from jax import numpy as jnp
 import qujax
 
[…]
